fix(web): standardize responsive pagination controls
